backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Real-time WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# AI Integration
openai==1.3.7

# CORS
django-cors-headers==4.3.1

# Environment Variables
python-dotenv==1.0.0

# Redis (for Channels layer)
redis==5.0.1

# Rate Limiting
django-ratelimit==4.1.0

# Utilities
Pillow==10.1.0
python-decouple==3.8
